improved script removed build before packaging VERSION is now properly propagated

+++ generate-package.sh
@@ -18,10 +18,19 @@ esac
 
 GENERATOR="${1:-${GENERATOR}}"
 
+if [ -z "${VERSION}" ]
+then
+   echo "Environment variable VERSION is not set" >&2
+   exit 1
+fi
+
+echo "Generating package for ${VERSION}" >&2
+
 if [ -d "package" ]
 then
    rm -rf package || exit 1
 fi
+
 
 mkdir package || exit 1
 cd package
